+++ app/controllers/admin/site_steps_controller.rb
@@ -52,10 +52,10 @@ class Admin::SiteStepsController < AdminController
       if step == 'contexts'
         @contexts = Context.all
       end
-      if step == 'style'
+      if step == 'template'
         SiteSetting.create_color_settings @site
       end
-      if step == 'settings'
+      if step == 'style'
         SiteSetting.create_additional_settings @site
         gon.global.color_controller_id = COLOR_CONTROLLER_ID
         gon.global.color_controller_name = COLOR_CONTROLLER_NAME
@@ -67,6 +67,12 @@ class Admin::SiteStepsController < AdminController
         @main_image = @site.site_settings.where(name: 'main_image').first
         @alternative_image = @site.site_settings.where(name: 'alternative_image').first
         @favico = @site.site_settings.where(name: 'favico').first
+      end
+      if step == 'settings'
+        SiteSetting.create_site_settings @site
+        @translate_english = @site.site_settings.where(name: 'translate_english').first
+        @translate_spanish = @site.site_settings.where(name: 'translate_spanish').first
+        @translate_french = @site.site_settings.where(name: 'translate_french').first
       end
     end
     render_wizard
@@ -165,8 +171,45 @@ class Admin::SiteStepsController < AdminController
           end
         end
 
-      when 'style'
+      when 'template'
         @site = current_site
+        if save_button?
+          if @site.save
+            delete_session_key(:site, @site_id)
+            redirect_to admin_sites_path, notice: 'The site\'s main color might take a few minutes to be visible'
+          else
+            render_wizard
+          end
+        else
+          @site.form_step = 'template'
+
+          if @site.valid?
+            redirect_to next_wizard_path
+          else
+            render_wizard
+          end
+        end
+
+
+      # In this step, the site is always saved
+      when 'style'
+        settings = site_params.to_h
+        @site = params[:site_slug] ? Site.find_by(slug: params[:site_slug]) : Site.new(session[:site][@site_id])
+
+        begin
+          # If the user is editing
+          if @site.id
+            @site.site_settings.each do |site_setting|
+              setting = settings[:site_settings_attributes].values.select { |s| s['id'] == site_setting.id.to_s }
+              site_setting.assign_attributes setting.first.except('id', 'position', 'name') if setting.any?
+            end
+            # If the user is creating a new site
+          else
+            settings[:site_settings_attributes].map { |s| @site.site_settings.build(s[1]) }
+            @site.form_step = 'style'
+          end
+        end
+
         if save_button?
           if @site.save
             delete_session_key(:site, @site_id)
@@ -180,12 +223,12 @@ class Admin::SiteStepsController < AdminController
           if @site.valid?
             redirect_to next_wizard_path
           else
+            color_array = @site.site_settings.select{ |s| s.name == 'flag'}.first
+            gon.global.color_array = color_array[:value].split(' ').map { |x| {color: x} } if color_array
             render_wizard
           end
         end
 
-
-      # In this step, the site is always saved
       when 'settings'
         settings = site_params.to_h
         @site = params[:site_slug] ? Site.find_by(slug: params[:site_slug]) : Site.new(session[:site][@site_id])
@@ -206,7 +249,7 @@ class Admin::SiteStepsController < AdminController
 
         if @site.save
           delete_session_key(:site, @site_id)
-          redirect_to next_wizard_path, notice: 'The site\'s main color might take a few minutes to be visible'
+          redirect_to next_wizard_path(site_slug: @site.slug), notice: 'The site\'s main color might take a few minutes to be visible'
         else
           render_wizard
         end
@@ -227,7 +270,9 @@ class Admin::SiteStepsController < AdminController
         routes_attributes: [:host, :id],
         site_settings_attributes: [
           :id, :position, :value, :name, :image,
-          :attribution_link, :attribution_label
+          :attribution_link, :attribution_label,
+          :translate_english, :translate_french,
+          :translate_spanish, :pre_footer, :analytics_key, :keywords, :contact_email_address
         ]
       )
   end